DROTH-3097 New unit tests to own spec file

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/service/lane/LaneWorkListService.scala
@@ -16,18 +16,17 @@ class LaneWorkListService {
     }
   }
 
-  def insertToLaneWorkList(propertyName: String, optionalExistingValue: Option[Int], linkProperty: LinkProperties, vvhRoadLink: VVHRoadlink, username: Option[String]): Unit = {
-    propertyName match {
+  def insertToLaneWorkList(propertyName: String, optionalExistingValue: Option[Int], linkProperty: LinkProperties,
+                           vvhRoadLink: VVHRoadlink, username: Option[String], newTransaction: Boolean = true): Unit = {
+    val itemToInsert = propertyName match {
       case "traffic_direction" =>
         val newValue = linkProperty.trafficDirection.value
         val oldValue = optionalExistingValue.getOrElse(vvhRoadLink.trafficDirection.value)
         val timeStamp = DateTime.now()
         val createdBy = username.getOrElse("")
         val itemToInsert = LaneWorkListItem(0, vvhRoadLink.linkId, propertyName, oldValue, newValue, timeStamp, createdBy)
-        if(newValue != oldValue) {
-          PostGISDatabase.withDynTransaction {
-            workListDao.insertItem(itemToInsert)}
-        }
+        if(newValue != oldValue) Some(itemToInsert)
+        else None
       case "link_type" =>
         val newValue = linkProperty.linkType.value
         val oldValue = optionalExistingValue.getOrElse(99)
@@ -36,19 +35,28 @@ class LaneWorkListService {
         val itemToInsert = LaneWorkListItem(0, vvhRoadLink.linkId, propertyName, oldValue, newValue, timeStamp, createdBy)
 
         val twoWayLaneLinkTypeChange = twoWayLanes.map(_.value).contains(newValue) || twoWayLanes.map(_.value).contains(oldValue)
-        if(twoWayLaneLinkTypeChange && (newValue != oldValue)) {
-          PostGISDatabase.withDynTransaction {
-            workListDao.insertItem(itemToInsert)
-          }
+        if(twoWayLaneLinkTypeChange && (newValue != oldValue)) Some(itemToInsert)
+        else None
+      case _ => None
+    }
+
+    if (itemToInsert.isDefined) {
+      if (newTransaction) {
+        PostGISDatabase.withDynTransaction {
+          workListDao.insertItem(itemToInsert.get)
         }
-      case _ =>
+      }
+      else workListDao.insertItem(itemToInsert.get)
     }
   }
 
-  def deleteFromLaneWorkList(itemsToDelete: Set[Long]): Set[Long] = {
-    PostGISDatabase.withDynTransaction {
-      workListDao.deleteItemsById(itemsToDelete)
+  def deleteFromLaneWorkList(itemsToDelete: Set[Long], newTransaction: Boolean = true): Set[Long] = {
+    if (newTransaction) {
+      PostGISDatabase.withDynTransaction {
+        workListDao.deleteItemsById(itemsToDelete)
+      }
     }
+    else workListDao.deleteItemsById(itemsToDelete)
     itemsToDelete
   }
 }
